README.md
Description
This project involves the detailed analysis of four Udemy subjects with different courses. Four files, one for each subject, were presented in excel format with errors embedded in them. The incorrect, incomplete, duplicated, and improperly formatted data was removed or modified. This is to ensure that the results are reliable and of high quality.

This report is to convey with a clear purpose the analyzed data. How the data was obtained, the process took to clean and transform the data into usable charts and tables to derive insights from them for decision-making. Also, to recommend a course of action for the problem of increasing revenue. The most popular courses, highest-paid courses, the number of subscribers, ratings, level, and income generated from each level and course must be known to increase revenue.

The results achieved were strongly inclined towards a direction. The subject of interest has a 67.9% subscription rate, a total price of $92,675, and the highest subject rating. Although Web Development has the highest subscription rate (85%) for the top 20 subscribed subjects, the courses are mostly free.

Conclusion
Web development emerged from the Top 20 Subscribed courses as the most subscribed with an 85% subscription rate. Although it is the most subscribed, they are primarily free. Also, an increase in the number of lectures will increase revenue for each subject, provided the number of subscriptions also increases. This holds only if the number of subscriptions to paid courses also increases. To increase revenue, I recommend that the firm;

1. Increase the price of courses for Web development and Business Finance
2. Allocate money to producing more paid lectures for all, beginner and intermediate level courses for Web Development and Business Finance
3. Convert most free courses for Web Development and Business Finance to paid
4. Little to no resources should be expended on expert-level courses, or it should be out of the business investment plan.